+++ lib/logd.dart
@@ -3,14 +3,6 @@ library;
 export 'src/handler/handler.dart';
 export 'src/logger/logger.dart' hide LogBuffer, LogEntry;
 export 'src/stack_trace/stack_trace.dart' hide CallbackInfo;
-export 'src/time/time.dart'
-    hide
-        LocalTime,
-        Time,
-        TimezoneOffset,
-        abbreviatedMonthNames,
-        abbreviatedWeekdayNames,
-        commonDSTRules,
-        commonTimezones,
-        monthNames,
-        weekdayNames;
+export 'src/time/timestamp.dart';
+export 'src/time/timezone.dart'
+    hide DSTZoneRule, LocalTime, OffsetLiteralExt, TimezoneOffset;